+++ aofl-js-packages/router/modules/router-instance.js
@@ -4,7 +4,7 @@
  * @since 1.0.0
  * @author Arian Khosravi<arian.khosravi@aofl.com>
  */
-import Router from './Router';
+import {Router} from './Router';
 /**
  * Persistent instance of Router.
  *