+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Removed]
 ### [Fixed]
 - @aofl/picture - img size updates
+- @aofl/map-state-properties-mixin - unit tests
 ### [Security]
 
 ---